+++ editor.planx.uk/src/pages/FlowEditor/components/Sidebar/Publish/PublishDialog.tsx
@@ -1,132 +1,189 @@
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
-import Collapse from "@mui/material/Collapse";
-import Divider from "@mui/material/Divider";
-import List from "@mui/material/List";
-import ListItem from "@mui/material/ListItem";
+import Dialog from "@mui/material/Dialog";
+import DialogActions from "@mui/material/DialogActions";
+import DialogContent from "@mui/material/DialogContent";
+import DialogTitle from "@mui/material/DialogTitle";
+import Link from "@mui/material/Link";
+import Step from "@mui/material/Step";
+import StepLabel from "@mui/material/StepLabel";
+import Stepper from "@mui/material/Stepper";
 import Typography from "@mui/material/Typography";
-import { ComponentType as TYPES } from "@opensystemslab/planx-core/types";
 import React, { useState } from "react";
-import Caret from "ui/icons/Caret";
+import Input from "ui/shared/Input/Input";
 
-import {
-  AlteredExternalPortalsSummary,
-  ExternalPortal,
-} from "./AlteredExternalPortals";
-import { AlteredNode, AlteredNodeListItem } from "./AlteredNodes";
+import { AlteredNode, AlteredNodesSummaryContent } from "./AlteredNodes";
+import { ValidationCheck, ValidationChecks } from "./ValidationChecks";
 
-interface AlteredNodesSummary {
-  title: string;
-  portals: ExternalPortal[];
-  updated: number;
-  deleted: number;
+interface NoChangesDialogProps {
+  dialogOpen: boolean;
+  setDialogOpen: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
-export const AlteredNodesSummaryContent = (props: {
-  alteredNodes: AlteredNode[];
-  lastPublishedTitle: string;
-}) => {
-  const { alteredNodes, lastPublishedTitle } = props;
-  const [expandNodes, setExpandNodes] = useState<boolean>(false);
-
-  const changeSummary: AlteredNodesSummary = {
-    title: lastPublishedTitle,
-    portals: [],
-    updated: 0,
-    deleted: 0,
-  };
-
-  alteredNodes.map((node) => {
-    if (node.id === "0") {
-      changeSummary["title"] =
-        "You are publishing the main service for the first time.";
-    } else if (node.id && Object.keys(node).length === 1) {
-      changeSummary["deleted"] += 1;
-    } else if (node.type === TYPES.InternalPortal) {
-      if (node.data?.text?.includes("/") && !node.data?.text?.includes(" ")) {
-        changeSummary["portals"].push({ ...node.data, flowId: node.id });
-      }
-    } else if (node.type) {
-      changeSummary["updated"] += 1;
-    }
-
-    return changeSummary;
-  });
+export const NoChangesDialog = (props: NoChangesDialogProps) => {
+  const { dialogOpen, setDialogOpen } = props;
 
   return (
-    <Box pb={2}>
-      <Typography variant="h4" component="h3" gutterBottom>
-        {`Changes`}
-      </Typography>
-      {changeSummary["title"] && (
-        <Typography variant="body2">{changeSummary["title"]}</Typography>
-      )}
-      {(changeSummary["updated"] > 0 || changeSummary["deleted"] > 0) && (
-        <Box pb={2}>
-          <List sx={{ listStyleType: "disc", marginLeft: 3 }}>
-            <ListItem
-              key={"updated"}
-              disablePadding
-              sx={{ display: "list-item" }}
-            >
-              <Typography variant="body2">{`${changeSummary["updated"]} nodes have been updated or added`}</Typography>
-            </ListItem>
-            <ListItem
-              key={"deleted"}
-              disablePadding
-              sx={{ display: "list-item" }}
-            >
-              <Typography variant="body2">{`${changeSummary["deleted"]} nodes have been deleted`}</Typography>
-            </ListItem>
-          </List>
-          <Box
-            style={{
-              display: "flex",
-              flexDirection: "column",
-              marginBottom: 2,
-            }}
-          >
-            <Box
-              style={{
-                display: "flex",
-                flexDirection: "row",
-                alignItems: "center",
-              }}
-            >
-              <Typography variant="body2">{`See detailed changelog `}</Typography>
-              <Button
-                onClick={() => setExpandNodes((expandNodes) => !expandNodes)}
-                size="small"
-                disableRipple
-              >
-                <Caret
-                  expanded={expandNodes}
-                  color="primary"
-                  titleAccess={
-                    expandNodes ? "Less information" : "More information"
-                  }
-                />
-              </Button>
-            </Box>
-            <Collapse in={expandNodes}>
-              <Box pb={1}>
-                <ul>
-                  {alteredNodes.map((node) => (
-                    <AlteredNodeListItem node={node} />
-                  ))}
-                </ul>
-              </Box>
-            </Collapse>
-          </Box>
-        </Box>
-      )}
-      <Divider />
-      {changeSummary["portals"].length > 0 && (
-        <>
-          <AlteredExternalPortalsSummary portals={changeSummary["portals"]} />
-          <Divider />
-        </>
-      )}
-    </Box>
+    <Dialog
+      open={dialogOpen}
+      onClose={() => setDialogOpen(false)}
+      aria-labelledby="alert-dialog-title"
+      aria-describedby="alert-dialog-description"
+      maxWidth="md"
+    >
+      <DialogTitle variant="h3" component="h1">
+        {`Check for changes to publish`}
+      </DialogTitle>
+      <DialogContent>
+        <Typography variant="body2">{`No new changes to publish`}</Typography>
+      </DialogContent>
+      <DialogActions sx={{ paddingX: 2 }}>
+        <Button onClick={() => setDialogOpen(false)}>KEEP EDITING</Button>
+        <Button color="primary" variant="contained" disabled={true}>
+          PUBLISH
+        </Button>
+      </DialogActions>
+    </Dialog>
+  );
+};
+
+interface ChangesDialogProps {
+  dialogOpen: boolean;
+  setDialogOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  alteredNodes: AlteredNode[];
+  lastPublishedTitle: string;
+  validationChecks: ValidationCheck[];
+  previewURL: string;
+  summary?: string;
+  setSummary: React.Dispatch<React.SetStateAction<string | undefined>>;
+  handlePublish: () => Promise<void>;
+}
+
+export const ChangesDialog = (props: ChangesDialogProps) => {
+  const {
+    dialogOpen,
+    setDialogOpen,
+    alteredNodes,
+    lastPublishedTitle,
+    validationChecks,
+    previewURL,
+    summary,
+    setSummary,
+    handlePublish,
+  } = props;
+
+  const steps = ["Review", "Test", "Publish"];
+
+  const [activeStep, setActiveStep] = useState<number>(0);
+
+  const handleNext = () => {
+    setActiveStep(activeStep + 1);
+  };
+
+  const handleBack = () => {
+    setActiveStep(activeStep - 1);
+  };
+
+  const ReviewStep = () => {
+    return (
+      <>
+        <DialogTitle variant="h3" component="h1">
+          {`Review`}
+        </DialogTitle>
+        <DialogContent>
+          <>
+            <AlteredNodesSummaryContent
+              alteredNodes={alteredNodes}
+              lastPublishedTitle={lastPublishedTitle}
+            />
+            <ValidationChecks validationChecks={validationChecks} />
+          </>
+        </DialogContent>
+        <DialogActions sx={{ paddingX: 2 }}>
+          <Button onClick={() => setDialogOpen(false)}>KEEP EDITING</Button>
+          <Button color="primary" variant="contained" onClick={handleNext}>
+            NEXT
+          </Button>
+        </DialogActions>
+      </>
+    );
+  };
+
+  const TestStep = () => {
+    return (
+      <>
+        <DialogTitle variant="h3" component="h1">
+          {`Test`}
+        </DialogTitle>
+        <DialogContent>
+          <Typography variant="body2">
+            {`Preview these content changes in-service before publishing `}
+            <Link href={previewURL} target="_blank">
+              {`here (opens in a new tab).`}
+            </Link>
+          </Typography>
+        </DialogContent>
+        <DialogActions sx={{ paddingX: 2 }}>
+          <Button onClick={handleBack}>BACK</Button>
+          <Button color="primary" variant="contained" onClick={handleNext}>
+            NEXT
+          </Button>
+        </DialogActions>
+      </>
+    );
+  };
+
+  const PublishStep = () => {
+    return (
+      <>
+        <DialogTitle variant="h3" component="h1">
+          {`Publish`}
+        </DialogTitle>
+        <DialogContent>
+          <Input
+            bordered
+            required
+            type="text"
+            name="summary"
+            value={summary || ""}
+            placeholder="Summarise your changes..."
+            onChange={(e) => setSummary(e.target.value)}
+          />
+        </DialogContent>
+        <DialogActions sx={{ paddingX: 2 }}>
+          <Button onClick={handleBack}>BACK</Button>
+          <Button color="primary" variant="contained" onClick={handlePublish}>
+            PUBLISH
+          </Button>
+        </DialogActions>
+      </>
+    );
+  };
+
+  return (
+    <Dialog
+      open={dialogOpen}
+      onClose={() => setDialogOpen(false)}
+      aria-labelledby="alert-dialog-title"
+      aria-describedby="alert-dialog-description"
+      maxWidth="md"
+    >
+      <Box padding={2}>
+        <Stepper activeStep={activeStep}>
+          {steps.map((label) => (
+            <Step key={label}>
+              <StepLabel>{label}</StepLabel>
+            </Step>
+          ))}
+        </Stepper>
+        {
+          {
+            0: <ReviewStep />,
+            1: <TestStep />,
+            2: <PublishStep />,
+          }[activeStep]
+        }
+      </Box>
+    </Dialog>
   );
 };
