+++ editor.planx.uk/src/pages/FlowEditor/components/Sidebar/Publish/PublishFlowButton.tsx
@@ -1,10 +1,5 @@
 import Box from "@mui/material/Box";
 import Button from "@mui/material/Button";
-import Dialog from "@mui/material/Dialog";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
-import DialogTitle from "@mui/material/DialogTitle";
-import Link from "@mui/material/Link";
 import Typography from "@mui/material/Typography";
 import { logger } from "airbrake";
 import { AxiosError } from "axios";
@@ -12,11 +7,10 @@ import { useStore } from "pages/FlowEditor/lib/store";
 import { formatLastPublishMessage } from "pages/FlowEditor/utils";
 import React, { useState } from "react";
 import { useAsync } from "react-use";
-import Input from "ui/shared/Input/Input";
 
 import { AlteredNode } from "./AlteredNodes";
-import { AlteredNodesSummaryContent } from "./PublishDialog";
-import { ValidationCheck, ValidationChecks } from "./ValidationChecks";
+import { ChangesDialog, NoChangesDialog } from "./PublishDialog";
+import { ValidationCheck } from "./ValidationChecks";
 
 export const PublishFlowButton: React.FC<{ previewURL: string }> = ({
   previewURL,
@@ -111,63 +105,24 @@ export const PublishFlowButton: React.FC<{ previewURL: string }> = ({
         >
           CHECK FOR CHANGES TO PUBLISH
         </Button>
-        <Dialog
-          open={dialogOpen}
-          onClose={() => setDialogOpen(false)}
-          aria-labelledby="alert-dialog-title"
-          aria-describedby="alert-dialog-description"
-          maxWidth="md"
-        >
-          <DialogTitle variant="h3" component="h1">
-            {`Check for changes to publish`}
-          </DialogTitle>
-          <DialogContent>
-            {alteredNodes?.length ? (
-              <>
-                <AlteredNodesSummaryContent
-                  alteredNodes={alteredNodes}
-                  lastPublishedTitle={lastPublishedTitle}
-                />
-                <ValidationChecks validationChecks={validationChecks} />
-                <Box pb={2}>
-                  <Typography variant="body2">
-                    {`Preview these content changes in-service before publishing `}
-                    <Link href={previewURL} target="_blank">
-                      {`here (opens in a new tab).`}
-                    </Link>
-                  </Typography>
-                </Box>
-                <Input
-                  bordered
-                  type="text"
-                  name="summary"
-                  value={summary || ""}
-                  placeholder="Summarise your changes..."
-                  onChange={(e) => setSummary(e.target.value)}
-                />
-              </>
-            ) : (
-              <Typography variant="body2">
-                {`No new changes to publish`}
-              </Typography>
-            )}
-          </DialogContent>
-          <DialogActions sx={{ paddingX: 2 }}>
-            <Button onClick={() => setDialogOpen(false)}>KEEP EDITING</Button>
-            <Button
-              color="primary"
-              variant="contained"
-              onClick={handlePublish}
-              disabled={
-                !alteredNodes ||
-                alteredNodes.length === 0 ||
-                validationChecks.filter((v) => v.status === "Fail").length > 0
-              }
-            >
-              PUBLISH
-            </Button>
-          </DialogActions>
-        </Dialog>
+        {!alteredNodes || alteredNodes?.length === 0 ? (
+          <NoChangesDialog
+            dialogOpen={dialogOpen}
+            setDialogOpen={setDialogOpen}
+          />
+        ) : (
+          <ChangesDialog
+            dialogOpen={dialogOpen}
+            setDialogOpen={setDialogOpen}
+            alteredNodes={alteredNodes}
+            lastPublishedTitle={lastPublishedTitle}
+            validationChecks={validationChecks}
+            previewURL={previewURL}
+            summary={summary}
+            setSummary={setSummary}
+            handlePublish={handlePublish}
+          />
+        )}
         <Box mr={0}>
           <Typography variant="caption">{lastPublishedTitle}</Typography>
         </Box>
